refactor(FlatDB): convert HSST builder/reader offsets from ulong to long
Mechanical follow-up to the long-finality widening work. Offset/length
values inside the HSST builder and B-tree reader were ulong; convert
them to long for consistency with the rest of the long-finality code.
Cast at the BSearchIndex boundary (BaseOffset stays ulong as it is
shared infrastructure).

- HsstBuilder.HsstEntry.MetadataStart and the FinishValueWrite local
  (drop the redundant checked((ulong)...) cast on a positive long).
- HsstIndexBuilder: NodeInfo.ChildOffset and all baseOffset / minVal /
  maxVal / minOff / maxOff / childOffset locals; widen MinBytesFor and
  WriteUInt64LE helper signatures.
- HsstBTreeReader: childOffset / metaStart locals, drop now-redundant
  (long) casts at the read sites.

Bloom keys, MemoryMarshal.Read<ulong> hash reads, bit-hack midpoints,
overflow-safe bounds checks, and binary u64 writes remain ulong.

+++ src/Nethermind/Nethermind.State.Flat/Hsst/HsstBTreeReader.cs
@@ -41,10 +41,10 @@ internal static class HsstBTreeReader
                 {
                     if (!node.TryGetFloor(key, out _, out ReadOnlySpan<byte> childValueBytes))
                         return false;
-                    ulong childOffset = BSearchIndex.BSearchIndexReader.ReadUInt64LE(childValueBytes) + node.Metadata.BaseOffset;
+                    long childOffset = (long)(BSearchIndex.BSearchIndexReader.ReadUInt64LE(childValueBytes) + node.Metadata.BaseOffset);
                     // childOffset is the inclusive last byte of the child node (0-indexed within the HSST).
                     // Exclusive end in reader-absolute terms = bound.Offset + childOffset + 1.
-                    currentAbsEnd = bound.Offset + (long)childOffset + 1;
+                    currentAbsEnd = bound.Offset + childOffset + 1;
                     continue;
                 }
 
@@ -60,8 +60,8 @@ internal static class HsstBTreeReader
                     if (!key.StartsWith(p) || !key[p.Length..].StartsWith(separator)) return false;
                 }
 
-                ulong metaStart = BSearchIndex.BSearchIndexReader.ReadUInt64LE(metaBytes) + node.Metadata.BaseOffset;
-                long absMetaStart = bound.Offset + (long)metaStart;
+                long metaStart = (long)(BSearchIndex.BSearchIndexReader.ReadUInt64LE(metaBytes) + node.Metadata.BaseOffset);
+                long absMetaStart = bound.Offset + metaStart;
 
                 // Read up to 11 bytes from absMetaStart: enough for ValueLength (≤10
                 // for long LEB128) + KeyLength (1 byte). KeyLength only consumed when
